push ulang karena tadi tidak lengkap

+++ exepemdas3_157/exepemdas3_157.cpp
@@ -3,15 +3,15 @@ using namespace std;
 
 class bidangDatar {
 private:
-	double x;														// variabel untuk menyimpan input dari lingkaran maupun bujursangkar
+	int x;														// variabel untuk menyimpan input dari lingkaran maupun bujursangkar
 public:
 	bidangDatar() {												// constructor
 		x = 0;
 	}
 	virtual void input() {}										// fungsi yang menerima input dan mengirim input melalui fungsi setX untuk disimpan di x
-	virtual float Luas(double a) { return 0; }						// fungsi untuk menghitung luas
-	virtual float keliling(double a) { return 0; }					// fungsi untuk menghitung keliling
-	void setX(double a) {											// fungsi untuk memberi/mengirim nilai pada x
+	virtual float Luas(int a) { return 0; }						// fungsi untuk menghitung luas
+	virtual float keliling(int a) { return 0; }					// fungsi untuk menghitung keliling
+	void setX(int a) {											// fungsi untuk memberi/mengirim nilai pada x
 		this->x = a;
 	}
 	int getX() {												// fungsi untuk membaca/mengambil nilai dalam x
@@ -23,18 +23,18 @@ public:															// membuat akses modifier public
 	void input() {												// menambahkan metod input
 		cout << "Bangun Datar Lingkaran telah dibuat";			// menampilkan bahwa progam lingkaran telah dibuat
 		cout << "\nMasukkan Jejari Lingkaran  : ";				// menampilkan jumlah jejari yang akan dimasukkan
-		double r;													// r = jejari
+		int r;													// r = jejari
 		cin >> r;												// input jejari yang akan dimasukkan
 		setX(r);												// masukkan fungsi get(x) / mencari x berdasarkan r = jejari
 	}
 
 	float Luas() {												// membuat fungsi luas
-		double r = getX();											// r = fungsi get x
+		int r = getX();											// r = fungsi get x
 		return 3.14 * r * r;									// mengembalikan fungsi get(x) menjadi rumus luas lingkaran = pi x r^2
 	}
 
 	float Keliling() {											// membuat fungsi keliling
-		double r = getX();											// r = fungsi get(x)
+		int r = getX();											// r = fungsi get(x)
 		return 2 * 3.14 * r;									// mengembalikan fungsi get(x) menjadi rumus keliling lingkaran = 2 x pi x r
 	}
 };
@@ -44,18 +44,18 @@ public:															// membuat modifier public
 	void input() {												// membuat fungsi input
 		cout << "\nBangun Datar Bujursangkar telah dibuat";		// menampilkan bahwa bujur sangkar telah dibuat
 		cout << "\nMasukkan Sisi Bujursangkar : ";				// menampilkan panjang sisi yang akan diinput
-		double s;													// s = sisi
+		int s;													// s = sisi
 		cin >> s;												// masukkan panjang sisi yang akan diinput
 		setX(s);												// masukkan fungsi get(x) / mencari x berdasarkan s = sisi
 	}
 
 	float Luas() {												// membuat fungsi luas
-		double s = getX();											// mencari x berdasarkan s = sisi
+		int s = getX();											// mencari x berdasarkan s = sisi
 		return s * s;											// mengembalikan fungsi get(x) menjadi rumus luas bujur sangkar s x s = sisi x sisi
 	}
 
 	float Keliling() {											// membuat fungsi keliling
-		double s = getX();											// mencari x berdasarkan s = sisi
+		int s = getX();											// mencari x berdasarkan s = sisi
 		return 4 * s;											// mengembalikan fungsi get(x) menjadi rumus keliling bujur sangkar 4 x s = 4 x sisi
 	}
 };
@@ -80,3 +80,5 @@ int main() { /*lengkapi disini*/
 
 	return 0;
 }
+
+
